requirements.txt
opencv-python
mediapipe
numpy
Pillow
# winsound is built-in on Windows
